+++ ensdart_v3/test/utils/normalize_test.dart
@@ -5,9 +5,9 @@ import 'package:test/test.dart';
 void main() {
   group('namehash', () {
     test("returns namehash for name", () async {
-      final nh = await nameHash('test.eth');
+      final nh = ENSNormalize();
       expect(
-          nh,
+          nh.nameHash('test.eth'),
           equals(
               "0xeb4f647bea6caa36333c816d7b46fdcb05f9466ecacc140ea8c66faf15b3d9f1"));
     }, timeout: Timeout.factor(2));