+++ ens_normalize/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.3
+
+- add escape to Normalization class
+- add isCombiningMark and shouldEscape functions
+
+## 1.0.2
+
+- Used Dart freezed for Normalization Dart class.
+- Parsed Spec.Json to b64 encoded string literal.
+
 ## 1.0.1
 
 - Renamed ENSnormalize method names.